+++ src/de_datalake_bulletin_dataload/definitions.py
@@ -12,7 +12,6 @@ defs = Definitions(
         ),
         "http_client": HTTPClientResource(
             user_agent=EnvVar("USER_AGENT"),
-            base_url=EnvVar("BULLETIN_WP_BASE_URL"),
             get_config=ConfigResource(
                 config_path=EnvVar("CONFIG_PATH")
             )